improve: regex
